+++ tests/maxo/routing/test_flags.py
@@ -1,7 +1,6 @@
 from typing import Any
 
 import pytest
-from magic_filter import F
 
 import maxo
 from maxo.routing.ctx import Ctx
@@ -12,13 +11,11 @@ from maxo.routing.flags import (
     Flag,
     FlagDecorator,
     FlagGenerator,
-    check_flags,
     extract_flags,
     extract_flags_from_object,
     flags,
     get_flag,
 )
-from maxo.routing.handlers.update import UpdateHandler
 from maxo.types import MessageCreated
 from tests.factories import make_flagged_handler
 
@@ -193,33 +190,3 @@ class TestGetFlag:
 
         assert get_flag(ctx, "spam") is None
         assert get_flag(ctx, "spam", default=42) == 42
-
-
-class TestCheckFlags:
-    @pytest.mark.parametrize(
-        ("handler_flags", "magic", "expected"),
-        [
-            ({}, F.test, None),
-            ({"test": True}, F.test, True),
-            ({"test": True}, F.spam, None),
-            ({"test": {"spam": "eggs"}}, F.test, {"spam": "eggs"}),
-        ],
-    )
-    def test_check_flags(
-        self,
-        handler_flags: dict[str, Any],
-        magic: Any,
-        expected: Any,
-    ) -> None:
-        ctx = Ctx({HANDLER_KEY: make_flagged_handler(**handler_flags)})
-
-        assert check_flags(ctx, magic) == expected
-
-    def test_nested_attribute_access(self) -> None:
-        @flags.chat_action(action="typing_on", interval=1)
-        async def func(update: MessageCreated) -> None:
-            pass
-
-        ctx = Ctx({HANDLER_KEY: UpdateHandler(func)})
-
-        assert check_flags(ctx, F.chat_action.action) == "typing_on"
